+++ packages/core/src/config/types.ts
@@ -159,11 +159,6 @@ export interface Config<P = APIEvent> {
   trackResources?: boolean;
 
   /**
-   * Track web vitals attribution data (default: true)
-   */
-  trackWebVitalsAttribution?: boolean;
-
-  /**
    * Configuration for the web vitals instrumentation
    */
   webVitalsInstrumentation?: {
@@ -181,8 +176,6 @@ export interface Config<P = APIEvent> {
 
     /**
      * Track web vitals attribution data (default: true)
-     *
-     * Functionally the same as setting `trackWebVitalsAttribution` to true.
      */
     trackAttribution?: boolean;
   };